.scripts/tags_manager/requirements.txt
oyaml
python-frontmatter
requests
